+++ components/CardFeatures/cardFeatures.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const cardFeatures = () => {
+  return <div>cardFeatures</div>;
+};
+
+export default cardFeatures;
