+++ components/Features/features.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const features = () => {
+  return (
+    <div>features</div>
+  )
+}
+
+export default features
