+++ components/Footer/footer.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const footer = () => {
+  return (
+    <div>footer</div>
+  )
+}
+
+export default footer
