+++ components/Hero/hero.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const hero = () => {
+  return (
+    <div>hero</div>
+  )
+}
+
+export default hero
