+++ components/ItemBenefits/itemBenefits.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const itemBenefits = () => {
+  return (
+    <div>itemBenefits</div>
+  )
+}
+
+export default itemBenefits
