+++ components/Navbar/nav.js
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const nav = () => {
+  return (
+    <div>nav</div>
+  )
+}
+
+export default nav
