--- pages/cardFeatures.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const cardFeatures = () => {
-  return (
-    <div>cardFeatures</div>
-  )
-}
-
-export default cardFeatures
