--- pages/features.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const features = () => {
-  return (
-    <div>features</div>
-  )
-}
-
-export default features
