--- pages/footer.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const footer = () => {
-  return (
-    <div>footer</div>
-  )
-}
-
-export default footer
