--- pages/hero.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const hero = () => {
-  return (
-    <div>hero</div>
-  )
-}
-
-export default hero
