--- pages/itemBenefits.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const itemBenefits = () => {
-  return (
-    <div>itemBenefits</div>
-  )
-}
-
-export default itemBenefits
