--- pages/nav.js
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const nav = () => {
-  return (
-    <div>nav</div>
-  )
-}
-
-export default nav
